requirements.txt
streamlit
numpy
pandas
Pillow
scipy
hmmlearn
scikit-learn
matplotlib
